requirements/dev.txt
-r base.txt
mypy
black
isort
autoflake
flake8
pytest
sqlalchemy-stubs
pytest-cov